protobuf-zip-imports/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
grpcio==1.65.4
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.65.4
    # via -r requirements.in
protobuf==5.27.3
    # via
    #   -r requirements.in
    #   grpcio-tools
setuptools==72.1.0
    # via grpcio-tools
